requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
pyodbc==5.0.1
netmiko==4.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
